+++ src/ts/Ads/Ads.ts
@@ -33,7 +33,7 @@ import { AdsConfigurationParser } from 'Ads/Parsers/AdsConfigurationParser';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 import { GameSessionCounters } from 'Ads/Utilities/GameSessionCounters';
 import { IosUtils } from 'Ads/Utilities/IosUtils';
-import { ProgrammaticTrackingService, ChinaMetric, ProgrammaticTrackingError, MiscellaneousMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
+import { ChinaMetric, ProgrammaticTrackingError, MiscellaneousMetric, LoadMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
 import { SdkStats } from 'Ads/Utilities/SdkStats';
 import { SessionDiagnostics } from 'Ads/Utilities/SessionDiagnostics';
 import { InterstitialWebPlayerContainer } from 'Ads/Utilities/WebPlayer/InterstitialWebPlayerContainer';
@@ -80,6 +80,10 @@ import { China } from 'China/China';
 import { IStore } from 'Store/IStore';
 import { RequestManager } from 'Core/Managers/RequestManager';
 import { AbstractAdUnitParametersFactory } from 'Ads/AdUnits/AdUnitParametersFactory';
+import { RefreshManager } from 'Ads/Managers/RefreshManager';
+import { PerPlacementLoadManager } from 'Ads/Managers/PerPlacementLoadManager';
+import { MediationMetaData } from 'Core/Models/MetaData/MediationMetaData';
+import { ZyngaLoadTest } from 'Core/Models/ABGroup';
 
 export class Ads implements IAds {
 
@@ -100,16 +104,13 @@ export class Ads implements IAds {
     public PlacementManager: PlacementManager;
     public AssetManager: AssetManager;
     public CampaignManager: CampaignManager;
-    public RefreshManager: CampaignRefreshManager;
+    public RefreshManager: RefreshManager;
 
     private static _forcedConsentUnit: boolean = false;
 
     private _currentAdUnit: AbstractAdUnit;
     private _showing: boolean = false;
-    private _creativeUrl?: string;
-    private _requestDelay: number;
-    private _wasRealtimePlacement: boolean = false;
-
+    private _loadApiEnabled: boolean = false;
     private _core: ICore;
     private _store: IStore;
 
@@ -171,10 +172,16 @@ export class Ads implements IAds {
 
             this.PlacementManager = new PlacementManager(this.Api, this.Config);
 
-            return this.PrivacyManager.getConsentAndUpdateConfiguration().catch(() => {
+            const promises = [];
+            if (CustomFeatures.isWhiteListedForLoadApi(this._core.ClientInfo.getGameId()) && ZyngaLoadTest.isValid(this._core.Config.getAbGroup())) {
+                promises.push(this.setupLoadApi());
+            }
+            promises.push(this.PrivacyManager.getConsentAndUpdateConfiguration().catch(() => {
                 // do nothing
                 // error happens when consent value is undefined
-            });
+            }));
+
+            return Promise.all(promises);
         }).then(() => {
             const defaultPlacement = this.Config.getDefaultPlacement();
             this.Api.Placement.setDefaultPlacement(defaultPlacement.getId());
@@ -240,8 +247,11 @@ export class Ads implements IAds {
             RequestManager.setAuctionProtocol(this._core.Config, this.Config, this._core.NativeBridge.getPlatform(), this._core.ClientInfo);
 
             this.CampaignManager = new CampaignManager(this._core.NativeBridge.getPlatform(), this._core, this._core.Config, this.Config, this.AssetManager, this.SessionManager, this.AdMobSignalFactory, this._core.RequestManager, this._core.ClientInfo, this._core.DeviceInfo, this._core.MetaDataManager, this._core.CacheBookkeeping, this.ContentTypeHandlerManager, this._core.JaegerManager, this.BackupCampaignManager);
-            this.RefreshManager = new CampaignRefreshManager(this._core.NativeBridge.getPlatform(), this._core.Api, this._core.Config, this.Api, this._core.WakeUpManager, this.CampaignManager, this.Config, this._core.FocusManager, this.SessionManager, this._core.ClientInfo, this._core.RequestManager, this._core.CacheManager);
-
+            if(this._loadApiEnabled) {
+                this.RefreshManager = new PerPlacementLoadManager(this._core.Api, this.Api, this.Config, this.CampaignManager, this._core.ClientInfo, this._core.FocusManager);
+            } else {
+                this.RefreshManager = new CampaignRefreshManager(this._core.NativeBridge.getPlatform(), this._core.Api, this._core.Config, this.Api, this._core.WakeUpManager, this.CampaignManager, this.Config, this._core.FocusManager, this.SessionManager, this._core.ClientInfo, this._core.RequestManager, this._core.CacheManager);
+            }
             SdkStats.initialize(this._core.Api, this._core.RequestManager, this._core.Config, this.Config, this.SessionManager, this.CampaignManager, this._core.MetaDataManager, this._core.ClientInfo, this._core.CacheManager);
 
             promo.initialize();
@@ -327,14 +337,8 @@ export class Ads implements IAds {
     public show(placementId: string, options: unknown, callback: INativeCallback): void {
         callback(CallbackStatus.OK);
         if (!this._core.FocusManager.isAppForeground()) {
-            if (CustomFeatures.shouldSampleAtTenPercent()) {
-                Diagnostics.trigger('ad_shown_in_background', {});
-            }
-
-            if (CustomFeatures.isShowingAdInBackground(this._core.ClientInfo.getGameId())) {
-                this._core.ProgrammaticTrackingService.reportMetric(MiscellaneousMetric.CampaignAttemptedToShowInBackground);
-                return;
-            }
+            this._core.ProgrammaticTrackingService.reportMetric(MiscellaneousMetric.CampaignAttemptedToShowInBackground);
+            return;
         }
 
         const campaign = this.RefreshManager.getCampaign(placementId);
@@ -361,7 +365,7 @@ export class Ads implements IAds {
 
             this._core.DeviceIdManager.getDeviceIds().then(() => {
                 Diagnostics.trigger('china_imei_collected', {
-                    imei: this._core.DeviceInfo instanceof AndroidDeviceInfo ? this._core.DeviceInfo.getDeviceId1(): "no-info"
+                    imei: this._core.DeviceInfo instanceof AndroidDeviceInfo ? this._core.DeviceInfo.getDeviceId1() : 'no-info'
                 });
             }).catch((error) => {
                 Diagnostics.trigger('china_imei_notcollected', error);
@@ -398,51 +402,20 @@ export class Ads implements IAds {
         }
 
         const trackingUrls = placement.getCurrentTrackingUrls();
-        if(trackingUrls) {
+        if (trackingUrls) {
             // Do not remove: Removing will currently break all tracking
             campaign.setTrackingUrls(trackingUrls);
+        } else {
+            this._core.ProgrammaticTrackingService.reportError(ProgrammaticTrackingError.MissingTrackingUrlsOnShow, contentType);
         }
 
         // First ad request within a game session can be made using recorded privacy information.
         // If game method has changed since, it should be reset before e.g. showing consent dialog
         this.resetOutdatedUserPrivacy();
 
-        if (placement.getRealtimeData() && !this.isConsentShowRequired()) {
-            this._core.Api.Sdk.logInfo('Unity Ads is requesting realtime fill for placement ' + placement.getId());
-            const start = Date.now();
-
-            const realtimeTimeoutInMillis = 1500;
-            Promises.withTimeout(this.CampaignManager.requestRealtime(placement, campaign.getSession()), realtimeTimeoutInMillis).then(realtimeCampaign => {
-                this._requestDelay = Date.now() - start;
-                this._core.Api.Sdk.logInfo(`Unity Ads received a realtime request in ${this._requestDelay} ms.`);
-
-                if(realtimeCampaign) {
-                    this._core.Api.Sdk.logInfo('Unity Ads received new fill for placement ' + placement.getId() + ', streaming new ad unit');
-                    this._wasRealtimePlacement = true;
-                    placement.setCurrentCampaign(realtimeCampaign);
-                    this.showAd(placement, realtimeCampaign, options);
-                } else {
-                    SessionDiagnostics.trigger('realtime_no_fill', {}, campaign.getSession());
-                    this._core.Api.Sdk.logInfo('Unity Ads received no new fill for placement ' + placement.getId() + ', opening old ad unit');
-                    this.showAd(placement, campaign, options);
-                }
-            }).catch((e) => {
-                if (e instanceof TimeoutError) {
-                    Diagnostics.trigger('realtime_network_timeout', {
-                        auctionId: campaign.getSession().getId()
-                    });
-                }
-                Diagnostics.trigger('realtime_error', {
-                    error: e
-                });
-                this._core.Api.Sdk.logInfo('Unity Ads realtime fill request for placement ' + placement.getId() + ' failed, opening old ad unit');
-                this.showAd(placement, campaign, options);
-            });
-        } else {
-            this.showConsentIfNeeded(options).then(() => {
-                this.showAd(placement, campaign, options);
-            });
-        }
+        this.showConsentIfNeeded(options).then(() => {
+            this.showAd(placement, campaign, options);
+        });
     }
 
     private resetOutdatedUserPrivacy() {
@@ -509,7 +482,7 @@ export class Ads implements IAds {
             const orientation = screenWidth >= screenHeight ? Orientation.LANDSCAPE : Orientation.PORTRAIT;
             AbstractPrivacy.createBuildInformation(this._core.NativeBridge.getPlatform(), this._core.ClientInfo, this._core.DeviceInfo, campaign, this._core.Config);
             this._currentAdUnit = this.getAdUnitFactory(campaign).create(campaign, placement, orientation, playerMetadataServerId || '', options);
-            this.RefreshManager.setCurrentAdUnit(this._currentAdUnit);
+            this.RefreshManager.setCurrentAdUnit(this._currentAdUnit, placement);
             if (this.Monetization.isInitialized()) {
                 this.Monetization.PlacementContentManager.setCurrentAdUnit(placement.getId(), this._currentAdUnit);
             }
@@ -540,22 +513,10 @@ export class Ads implements IAds {
             OperativeEventManager.setPreviousPlacementId(this.CampaignManager.getPreviousPlacementId());
             this.CampaignManager.setPreviousPlacementId(placement.getId());
 
-            // Temporary for realtime testing purposes
-            if (this._wasRealtimePlacement) {
-                this._currentAdUnit.onStart.subscribe(() => {
-                    const startDelay = Date.now() - start;
-                    Diagnostics.trigger('realtime_delay', {
-                        requestDelay: this._requestDelay,
-                        startDelay: startDelay,
-                        totalDelay: this._requestDelay + startDelay,
-                        auctionId: campaign.getSession().getId(),
-                        adUnitDescription: this._currentAdUnit.description()
-                    });
-                });
-            }
-            this._wasRealtimePlacement = false;
-
             this._currentAdUnit.show().then(() => {
+                if (this._loadApiEnabled) {
+                    this._core.ProgrammaticTrackingService.reportMetric(LoadMetric.LoadEnabledShow);
+                }
                 this.BackupCampaignManager.deleteBackupCampaigns();
             });
         });
@@ -576,6 +537,22 @@ export class Ads implements IAds {
 
     private onAdUnitClose(): void {
         this._showing = false;
+    }
+
+    private setupLoadApi(): Promise<void> {
+        this._loadApiEnabled = false;
+
+        return this._core.MetaDataManager.fetch(MediationMetaData).then((mediation) => {
+            if(mediation) {
+                const loadEnabled = mediation.isMetaDataLoadEnabled();
+                if(loadEnabled) {
+                    this._loadApiEnabled = true;
+                    this._core.ProgrammaticTrackingService.reportMetric(LoadMetric.LoadEnabledInitializationSuccess);
+                } else {
+                    this._core.ProgrammaticTrackingService.reportMetric(LoadMetric.LoadEnabledInitializationFailure);
+                }
+            }
+        });
     }
 
     private setupTestEnvironment(): void {
@@ -636,7 +613,7 @@ export class Ads implements IAds {
         }
 
         if(TestEnvironment.get('creativeUrl')) {
-            const creativeUrl = this._creativeUrl = TestEnvironment.get<string>('creativeUrl');
+            const creativeUrl = TestEnvironment.get<string>('creativeUrl');
             let response: string = '';
             const platform = this._core.NativeBridge.getPlatform();
             if(platform === Platform.ANDROID) {